feat: 优化调整界面，拆分各个模块为单独的一个配置页。
