+++ src/views/advantage/dialog.vue
@@ -38,6 +38,7 @@
 </template>
 
 <script setup lang="ts">
+// @ts-ignore
 import { ElMessage, UploadProps } from "element-plus";
 import { reactive, ref, defineEmits, defineExpose } from "vue";
 import { apiAddAdvantage, apiSetAdvantage } from "../../api";
